pyproject.toml
[project]
name = "jarvis"
version = "2.0.0"
description = "JARVIS — AI-first voice agent with local LLM, web search, and full OS control"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Voice
    "faster-whisper>=1.2.1",
    "sounddevice>=0.5.5",
    "scipy>=1.17.1",
    "pyttsx3>=2.99",
    # AI / LLM
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.31",
    "sentence-transformers>=3.0.0",
    # Memory
    "faiss-cpu>=1.9.0",
    "numpy>=2.4.4",
    # Web tools
    "duckduckgo-search>=8.1.1",
    "wikipedia>=1.4.0",
    "requests>=2.31.0",
    # System tools
    "pyautogui>=0.9.54",
    "psutil>=7.2.2",
    "pyperclip>=1.8.2",
    "keyboard>=0.13.5",
    "pynput>=1.8.1",
    # Environment
    "python-dotenv>=1.0.0",
    # UI
    "rich>=13.0.0",
]
